backend/pyproject.toml
[tool.poetry]
name = "jk-core"
version = "0.1.0"
description = "Курсовой проект Ягуара Ковалёва и Олега Шилова"
authors = ["Jaguar K. & Oleg Sh."]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.6.0"
djangorestframework-simplejwt = {extras = ["redis"], version = "^5.3.1"}
pillow = "^11.0.0"
django-filter = "^24.3"
python-decouple = "^3.8"
drf-yasg = "^1.21.8"
drf-spectacular = "^0.28.0"
django-redis = "^5.4.0"
gunicorn = "^23.0.0"
boto3 = "^1.35.94"
django-storages = "^1.14.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


